+++ src/pyatmo/const.py
@@ -10,6 +10,8 @@ ERRORS: dict[int, str] = {
     403: "Forbidden",
     404: "Not found",
     406: "Not Acceptable",
+    409: "Conflict",
+    429: "Too Many Requests",
     500: "Internal Server Error",
     502: "Bad Gateway",
     503: "Service Unavailable",